Also add mypy annotations

cc H-Huang awgu kwen2501 wanchaol fegin fduwjj wz337 d4l3k c-p-i-o

[ghstack-poisoned]

+++ torch/testing/_internal/common_distributed.py
@@ -29,6 +29,8 @@ import torch
 import torch._dynamo.test_case
 import torch.cuda.nccl
 import torch.distributed as c10d
+from torch._C._autograd import DeviceType
+from torch._C._distributed_c10d import _SymmetricMemory
 import torch.nn as nn
 from torch.testing._internal.common_utils import (
     FILE_SCHEMA,
@@ -337,6 +339,17 @@ def requires_mpi():
     return skip_but_pass_in_sandcastle_if(
         not c10d.is_mpi_available(),
         "c10d was not compiled with the MPI backend",
+    )
+
+
+def requires_multicast_support():
+    has_multicast_support = (
+        torch.cuda.is_available()
+        and _SymmetricMemory.has_multicast_support(DeviceType.CUDA, 0)
+    )
+    return skip_but_pass_in_sandcastle_if(
+        not has_multicast_support,
+        "multicast support is not available",
     )
 
 
@@ -671,7 +684,7 @@ class MultiProcessTestCase(TestCase):
         self.file_name = file_name
         self.run_test(test_name, parent_pipe)
 
-    def run_test(self, test_name: str, parent_pipe) -> None:
+    def run_test(self, test_name: str, parent_pipe, destroy_process_group=True) -> None:
         # Start event listener thread.
         signal_recv_pipe, signal_send_pipe = torch.multiprocessing.Pipe(duplex=False)
         event_listener_thread = threading.Thread(
@@ -714,13 +727,14 @@ class MultiProcessTestCase(TestCase):
             # Close pipe after done with test.
             parent_pipe.close()
 
-        try:
-            # Many test cases init a process group but do not destroy it.
-            # Some tests do destroy the pgs, and destroy can't be called twice.
-            # This avoids spewing warnings about improperly shutting down.
-            c10d.destroy_process_group()
-        except (AssertionError, ValueError):
-            pass
+        if destroy_process_group:
+            try:
+                # Many test cases init a process group but do not destroy it.
+                # Some tests do destroy the pgs, and destroy can't be called twice.
+                # This avoids spewing warnings about improperly shutting down.
+                c10d.destroy_process_group()
+            except (AssertionError, ValueError):
+                pass
 
     def _get_timedout_process_traceback(self) -> None:
         pipes = []
